Adding control and feedback code
Now in the file I have added everything that was previously stored in my folder in the citadel in hopes of making the knowledge more permanent. I need to go back and comment everything a little better, but right now there is code for the arduino to read the photodiode (voltage), and then matlab code for controlling the lenses. Eventually the arduino measurement will be used as a feedback mechanism